make it compatible with TPU
