+++ official/vision/beta/projects/centernet/configs/centernet.py
@@ -93,9 +93,6 @@ class SegmentationLoss(hyperparams.Config):
 class Losses(hyperparams.Config):
   detection: DetectionLoss = DetectionLoss()
   segmentation: SegmentationLoss = SegmentationLoss()
-  use_gaussian_bump: bool = True
-  use_odapi_gaussian: bool = False
-  gaussian_rad: int = -1
   gaussian_iou: float = 0.7
   class_offset: int = 1
